+++ game/entities/pois.h
@@ -32,9 +32,9 @@ class POIs
 
 public:
 
-    static void spawnDoor(const Vec2f &pos, uint16_t left, uint16_t top, uint8_t width, uint8_t height, POIType door);
+    static void spawnDoor(const Vec2i &pos, uint16_t left, uint16_t top, uint8_t width, uint8_t height, POIType door);
 
-    static void spawnShop(const Vec2f &pos, const Vec2f &jeep_loc, const Vec2f &boat_loc, const Vec2f &heli_loc);
+    static void spawnShop(const Vec2i &pos, const Vec2f &jeep_loc, const Vec2f &boat_loc, const Vec2f &heli_loc);
 
     static void unlockCurrent();
 
@@ -46,27 +46,27 @@ public:
 };
 
 template <int x1, int y1, int x2, int y2, int x3, int y3>
-void spawnShop(const Vec2f &loc) {
+void spawnShop(const Vec2i &loc) {
     POIs::spawnShop(loc,Vec2f(x1*6,y1*6),Vec2f(x2*6,y2*6),Vec2f(x3*6,y3*6));
 }
 
 template <int l, int t, int w, int h>
-void spawnDoorA(const Vec2f &loc) {
+void spawnDoorA(const Vec2i &loc) {
     POIs::spawnDoor(loc, l*6, t*6, w*6, h*6, POIType::DoorA);
 }
 
 template <int l, int t, int w, int h>
-void spawnDoorB(const Vec2f &loc) {
+void spawnDoorB(const Vec2i &loc) {
     POIs::spawnDoor(loc, l*6, t*6, w*6, h*6, POIType::DoorB);
 }
 
 template <int l, int t, int w, int h>
-void spawnDoorC(const Vec2f &loc) {
+void spawnDoorC(const Vec2i &loc) {
     POIs::spawnDoor(loc, l*6, t*6, w*6, h*6, POIType::DoorC);
 }
 
 template <int l, int t, int w, int h>
-void spawnDoorNone(const Vec2f &loc) {
+void spawnDoorNone(const Vec2i &loc) {
     POIs::spawnDoor(loc, l*6, t*6, w*6, h*6, POIType::DoorNone);
 }
